Controller recipes with test
