+++ src/mock/responseParam.js
@@ -1,0 +1,5 @@
+const res = {};
+res.status = jest.fn().mockReturnValue(res);
+res.json = jest.fn().mockReturnValue(res);
+
+module.exports = res;
