+++ src/mock/serviceGetRecipesPuppyReturn.js
@@ -1,0 +1,16 @@
+const mock = [
+  {
+    title: 'Cocoa pie',
+    href: 'http://yourrecipe.com',
+    ingredients: 'cocoa, milk, suggar',
+    thumbnail: '',
+  },
+  {
+    title: 'Cocoa pie',
+    href: 'http://yourrecipe.com',
+    ingredients: 'cocoa, milk, suggar',
+    thumbnail: '',
+  },
+];
+
+module.exports = mock;
