+++ src/mock/serviceGiphyReturn.js
@@ -1,0 +1,9 @@
+const mock = {
+  images: {
+    original: {
+      url: 'http://yourgiphy.com',
+    },
+  },
+};
+
+module.exports = mock;
