+++ src/services/getRecipesFromPuppy.js
@@ -1,20 +1,8 @@
 const axios = require('axios');
 
+const mock = require('../mock/serviceGetRecipesPuppyReturn');
+
 const isTest = process.env.NODE_ENV === 'test';
-const mock = [
-  {
-    title: 'Cocoa pie',
-    href: 'http://yourrecipe.com',
-    ingredients: 'cocoa, milk, suggar',
-    thumbnail: '',
-  },
-  {
-    title: 'Cocoa pie',
-    href: 'http://yourrecipe.com',
-    ingredients: 'cocoa, milk, suggar',
-    thumbnail: '',
-  },
-];
 
 module.exports = async (ingredients, page) => {
   if (isTest) return mock;
